+++ interfaces/PyInterface/gauss_jacobi_quad.py
@@ -12,7 +12,7 @@ Example::
 
 from __future__ import annotations
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 
 
 import ctypes